+++ cpp_tester/compiler.py
@@ -3,7 +3,7 @@ import yaml
 import os
 import subprocess
 
-from test_runner import TestRunner
+from classes.test_runner import TestRunner
 from util import make_test_output, get_executable_name
 
 class CppCompileTestRunner(TestRunner):
