+++ cpp_tester/googletest_runner.py
@@ -1,8 +1,8 @@
-from test_runner import TestRunner
+from classes.test_runner import TestRunner
 
 class GoogleTestRunner(TestRunner):
     def __init__(self):
         super().__init__()
-    
+
     def run_test(self):
         return {}
