+++ cpp_tester/linter.py
@@ -2,7 +2,7 @@ import json
 import os
 import subprocess
 
-from test_runner import TestRunner
+from classes.test_runner import TestRunner
 from util import make_test_output, get_executable_name
 
 class CppLinter(TestRunner):
